Refactor into packages
Packages Created:
dukeScripts
dukeScripts.Tasks
