+++ src/main/java/dukeScripts/Duke.java
@@ -1,4 +1,7 @@
-import java.util.ArrayList;
+package dukeScripts;
+
+import dukeScripts.Tasks.TDLTask;
+
 import java.util.Scanner;
 
 public class Duke {
