+++ src/main/java/dukeScripts/DukeCommandParser.java
@@ -1,0 +1,11 @@
+
+package dukeScripts;
+
+/**
+ * This class encapsulates the command parser for Duke which process the commands
+ * inputted into Duke by users.
+ */
+public class DukeCommandParser {
+
+
+}
