+++ src/main/java/dukeScripts/DukeStorageManager.java
@@ -1,0 +1,10 @@
+
+package dukeScripts;
+
+/**
+ * This class encapsulates the local storage functions of Duke.
+ */
+public class DukeStorageManager {
+
+
+}
