+++ src/main/java/dukeScripts/DukeUI.java
@@ -1,0 +1,9 @@
+package dukeScripts;
+
+/**
+ * This class encapsulates the interactions between Duke and the Command Line Interface.
+ */
+public class DukeUI {
+
+
+}
